requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
prometheus-fastapi-instrumentator
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
